requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorzero==2.0
fastapi==0.121.1
gpiozero==2.0.1
h11==0.16.0
idna==3.11
numpy==2.2.6
opencv-contrib-python==4.12.0.88
pigpio==1.78
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
